pyproject.toml
[project]
name = "mercati"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"fredapi>=0.5.2",
	"hydra-core>=1.3.2",
	"matplotlib>=3.10.0",
	"pandas>=2.3.0",
	"pandas-datareader>=0.10.0",
	"python-dotenv>=1.2.2",
	"torch>=2.10.0",
	"yfinance>=0.2.66",
]
